Remove install path in generated CSV
Install path was wrong because package is installed to relative path in
build system. This meant that generated CSV was invalid and pointed to
wrong location.
On the other hand there is no need to have absolute path there. User
accesses this file in same location as EULAs are so it is easy to locate
them and makes it independent on install path in this package.

This commit also reformats touched code to make it more readable (single
line to formated multiline).